Add tests and remove old resource scripts
